wonderwoman_facefusion_api/requirements.txt

fastapi
uvicorn
python-multipart
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
transformers
diffusers
accelerate
safetensors
Pillow
opencv-python
dlib
insightface==0.7.3
numpy==1.24.4
